requirements.txt
pytesseract
pandas
numpy
shapely
Pillow
lxml
scikit-learn
huggingface_hub
transformers
datasets
seqeval
torch
torchvision
accelerate
